api通信は子コンポーネントで行う
